Move top-level aggregate execution functions to own file
